+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Border on shipping preview
+- Focus on input
+
 ## [0.0.19] - 2020-09-10
 
 ### Fixed